base-requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mangum==0.19.0
python-dotenv==1.0.1
pydantic==2.9.2
rapidfuzz==3.9.7
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
faiss-cpu==1.8.0.post1
langchain==0.3.7
langchain-core==0.3.19
langgraph==0.2.39
langfuse==2.53.9
crewai==0.76.9
google-generativeai==0.8.3
langchain-google-genai==2.0.5
openai==1.52.2
langchain-openai==0.2.8
